+++ agent/discovery/src/lib.rs
@@ -1,5 +1,5 @@
-#[macro_use]
-extern crate error_chain;
+extern crate failure;
+extern crate failure_derive;
 extern crate serde;
 #[macro_use]
 extern crate serde_derive;
@@ -10,12 +10,14 @@ extern crate replicante_data_models;
 
 mod backends;
 mod config;
-mod errors;
+mod error;
 
 pub use self::backends::Iter;
 pub use self::backends::discover;
 pub use self::config::Config;
-pub use self::errors::*;
+pub use self::error::Error;
+pub use self::error::ErrorKind;
+pub use self::error::Result;
 
 // Expose internal models for validaion.
 pub use self::backends::DiscoveryFileModel;