+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 - Improve playgrounds usability.
+- Replace `error-chain` with `failure`.
 
 ## [0.2.0] - 2019-02-20
 ### Added